4. Token Sale

+++ contracts/FmhToken.sol
@@ -4,29 +4,29 @@ contract FmhToken {
 	string public name = 'FMH Token';
 	string public symbol = 'FMH';
 	string public standard = 'FMH Token v1.0';
-	uint256 public totalSupply;
+	uint public totalSupply;
 
 	event Transfer(
 		address indexed _from,
 		address indexed _to,
-		uint256 _value
+		uint _value
 	);
 
 	event Approval(
 		address indexed _owner,
 		address indexed _spender,
-		uint256 _value
+		uint _value
 	);
 	
-	mapping(address => uint256) public balanceOf;
-	mapping(address => mapping(address => uint256)) public allowance;
+	mapping(address => uint) public balanceOf;
+	mapping(address => mapping(address => uint)) public allowance;
 	
-	constructor(uint256 _initialSupply) public {
+	constructor(uint _initialSupply) public {
 		balanceOf[msg.sender] = _initialSupply;
 		totalSupply = _initialSupply;
 	}
 
-	function transfer(address _to, uint256 _value) public returns (bool success) {
+	function transfer(address _to, uint _value) public returns (bool success) {
         require(balanceOf[msg.sender] >= _value);
         
         balanceOf[msg.sender] -= _value;
@@ -37,7 +37,7 @@ contract FmhToken {
         return true;
 	}
 
-	function approve(address _spender, uint256 _value) public returns (bool success) {
+	function approve(address _spender, uint _value) public returns (bool success) {
 		allowance[msg.sender][_spender] = _value;
 
 		emit Approval(msg.sender, _spender, _value);
@@ -45,7 +45,7 @@ contract FmhToken {
 		return true;
 	}
 
-	function transferFrom(address _from, address _to, uint256 _value) public returns (bool success) {
+	function transferFrom(address _from, address _to, uint _value) public returns (bool success) {
 		require(_value <= balanceOf[_from]);
 		require(_value <= allowance[_from][msg.sender]);
